agentic-mirror/backend/embeddings.py
"""
Agentic Mirror — Embeddings + PCA Pipeline
See AGENT.md §9 — The Embeddings Pipeline

Uses OpenAI text-embedding-3-small for embedding generation.
PCA is the default dimensionality reduction (AGENT.md §14 Rule 9).
UMAP available as optional upgrade, not default.
"""

import os
import numpy as np
from openai import OpenAI
from sklearn.decomposition import PCA
from dotenv import load_dotenv

load_dotenv()

client = OpenAI(api_key=os.getenv("OPENAI_API_KEY"))


def embed_texts(texts: list[str]) -> list[list[float]]:
    """
    Generate embeddings for a list of text strings using OpenAI.

    Args:
        texts: List of sentences/phrases to embed.

    Returns:
        List of embedding vectors (each is a list of floats).

    See AGENT.md §9 — uses text-embedding-3-small model.
    """
    # TODO: Call client.embeddings.create(input=texts, model="text-embedding-3-small")
    # TODO: Extract and return [item.embedding for item in response.data]
    # TODO: Handle API errors gracefully (AGENT.md §14 Rule 10)
    pass


def reduce_to_2d(embeddings: list[list[float]]) -> list[tuple[float, float]]:
    """
    Reduce high-dimensional embeddings to 2D using PCA.

    Args:
        embeddings: List of embedding vectors from embed_texts().

    Returns:
        List of (x, y) tuples in 2D PCA space.

    See AGENT.md §14 Rule 9 — PCA first (fast, deterministic).
    """
    # TODO: Convert to numpy array
    # TODO: Fit PCA(n_components=2) and transform
    # TODO: Return as list of tuples
    pass


def reduce_to_2d_umap(embeddings: list[list[float]]) -> list[tuple[float, float]]:
    """
    Optional: Reduce to 2D using UMAP (slower, non-deterministic, better clusters).
    Not the default — see AGENT.md §14 Rule 9.

    Args:
        embeddings: List of embedding vectors.

    Returns:
        List of (x, y) tuples in 2D UMAP space.
    """
    # TODO: Import umap
    # TODO: Fit UMAP(n_components=2) and transform
    # TODO: Return as list of tuples
    pass


def label_axes(variance_ratio: np.ndarray, components: np.ndarray) -> dict[str, str]:
    """
    Name the PCA axes based on top-loading sentences.

    Args:
        variance_ratio: PCA explained variance ratio.
        components: PCA component loadings.

    Returns:
        {"x": "Risk vs Safety", "y": "Past vs Future"} (example).

    See AGENT.md §9 — uses a fast Claude call to name axes based on
    top-loading sentences in each principal component.
    """
    # TODO: Identify top-contributing dimensions for each component
    # TODO: Optionally call Claude to generate semantic axis labels
    # TODO: Return dict with "x" and "y" axis names
    pass
